PHX 1.5->1.7 upgrade
This was done by creating a new PHX 1.7 app, and coping our application
code into that app, and then deleting everything in the old project and
copying everything new in.

Currently the CSS is still not correct, but to create a save-point while
working on it we're going to do this massive upgrade commit and then
start working from there.

Several changes were made as part of this upgrade, including
renaming/updating the database runner script, and fixing some of the
bugs that cropped up from using the new elixir/erlang versions.
Some additional things, such as vendor/heroicons, have been comitted as
they are part of the default tailwind installation. Might as well keep
them for now and prune them later if needed.

This also migrated from SCSS to plain CSS. In practice it appears no
SCSS features were actually being used, so no big deal. But SCSS *is*
pretty useful, so it'd be interesting to migrate back someday.